requirements.txt
fastapi
uvicorn[standard]
groq
firebase-admin
pydantic
python-dotenv
streamlit
requests
httpx
